+++ src/main/kotlin/com/dh/baro/order/presentation/OrderController.kt
@@ -4,7 +4,7 @@ import com.dh.baro.core.Cursor
 import com.dh.baro.core.SliceResponse
 import com.dh.baro.core.auth.CurrentUser
 import com.dh.baro.order.application.OrderFacade
-import com.dh.baro.order.domain.OrderService
+import com.dh.baro.order.application.OrderQueryFacade
 import jakarta.validation.Valid
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.*
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.*
 @RestController
 @RequestMapping("/orders")
 class OrderController(
+    private val orderQueryFacade: OrderQueryFacade,
     private val orderFacade: OrderFacade,
-    private val orderService: OrderService,
 ) {
 
     @PostMapping
@@ -32,7 +32,7 @@ class OrderController(
         @CurrentUser userId: Long,
         @PathVariable orderId: Long,
     ): OrderDetailResponse =
-        OrderDetailResponse.from(orderService.getOrderDetail(userId, orderId))
+        OrderDetailResponse.from(orderQueryFacade.getOrderDetail(userId, orderId))
 
     @GetMapping
     @ResponseStatus(HttpStatus.OK)
@@ -41,7 +41,7 @@ class OrderController(
         @RequestParam(required = false) cursorId: Long?,
         @RequestParam(defaultValue = "10") size: Int,
     ): SliceResponse<OrderListItem> {
-        val slice = orderFacade.getOrdersByCursor(userId, cursorId, size)
+        val slice = orderQueryFacade.getOrdersByCursor(userId, cursorId, size)
         return SliceResponse.from(
             slice = slice,
             mapper = OrderListItem::from,
